chore: ignore index files in coverage (#8293)
