+++ vitest.config.ts
@@ -9,8 +9,12 @@ export default defineConfig({
 			all: true,
 			reporter: ['text', 'lcov', 'cobertura'],
 			include: ['src'],
-			// All ts files that only contain types, due to ALL
-			exclude: ['**/*.{interface,type,d}.ts'],
+			exclude: [
+				// All ts files that only contain types, due to ALL
+				'**/*.{interface,type,d}.ts',
+				// All index files that *should* only contain exports from other files
+				'**/index.{js,ts}',
+			],
 		},
 	},
 });
